Updating script to be more sequential and map planets and ships
Created a map_tic that will take the tic and map the location
of the ships and the planets that were conquered and lost
on that tic.

Also making the next call to map_tic happen after the success
of the previous call so it works in order as expected.